+++ CMakeLists.txt
@@ -59,12 +59,12 @@ juce_add_gui_app(GuiAppExample
 
 target_sources(GuiAppExample
     PRIVATE
-        Main.cpp
-        MainComponent.cpp
-        SoundProcessor.cpp
-        SoundProcessor.h
-        Wavetable.h
-        ErrorDef.h)
+        src/Main.cpp
+        src/MainComponent.cpp
+        src/SoundProcessor.cpp
+        src/SoundProcessor.h
+        src/Wavetable.h
+        src/ErrorDef.h)
 
 # `target_compile_definitions` adds some preprocessor definitions to our target. In a Projucer
 # project, these might be passed in the 'Preprocessor Definitions' field. JUCE modules also make use
